requirements.txt
Django==1.11.2
django-crispy-forms==1.6.1
django-filter==1.0.4
django-markdown-deux==1.0.5
django-multiselectfield==0.1.8
django-pagedown==0.1.3
djangorestframework==3.6.4
djangorestframework-jwt==1.11.0
gunicorn==19.7.1
Markdown==2.6.9
markdown2==2.3.4
olefile==0.44
Pillow==4.1.1
PyJWT==1.5.2
pytz==2017.2
